refactor(gui-v2): move jwt signup view to default signup page
