+++ packages/nc-gui-v2/pages/signup/[[token]]/none.vue
@@ -1,0 +1,6 @@
+<template>
+  <div class="flex items-center prose-2xl text-red-500 font-italic">
+    <MdiAlert />
+    Authentication not configured in configuration
+  </div>
+</template>
